Balance in search screen

+++ apps/wallet/instance/src/main/java/com/tonapps/tonkeeper/ui/screen/swap/search/list/all/TokenItem.kt
@@ -19,7 +19,7 @@ sealed class TokenItem(type: Int) : BaseListItem(type) {
         val contractAddress: String,
         val symbol: String,
         val name: String,
-        val balance: Float,
+        val balance: Double,
         val balanceFormat: CharSequence,
         val fiat: Float,
         val fiatFormat: CharSequence,
